fix(ui): make model requests owner-cancellable
